+++ Tests/Editor/YarnImporterTests.cs
@@ -352,7 +352,7 @@ position: 0,0
         }
 
         [Test]
-        public void YarnImporter_CanCreateYarnProgram()
+        public void YarnImporter_CanCreateYarnProject()
         {
 
             string scriptPath = "NewYarnScript.yarn";
@@ -365,21 +365,21 @@ position: 0,0
             var scriptImporter = AssetImporter.GetAtPath(scriptFullPath) as YarnImporter;
             
             // The script has no destination program after being created
-            Assert.Null(scriptImporter.DestinationProgram);
+            Assert.Null(scriptImporter.DestinationProject);
 
-            // Create a new Yarn Program for this script
-            var programPath = YarnImporterUtility.CreateYarnProgram(scriptImporter);
+            // Create a new Yarn Project for this script
+            var programPath = YarnImporterUtility.CreateYarnProject(scriptImporter);
             createdFilePaths.Add(programPath);
 
-            // The script now has a destination program
-            Assert.NotNull(scriptImporter.DestinationProgram);
+            // The script now has a destination project
+            Assert.NotNull(scriptImporter.DestinationProject);
 
-            var programImporter = AssetImporter.GetAtPath(programPath) as YarnProgramImporter;
+            var projectImporter = AssetImporter.GetAtPath(programPath) as YarnProjectImporter;
 
             var scriptTextAsset = AssetDatabase.LoadAssetAtPath<TextAsset>(scriptFullPath);
 
-            // The program includes the script importer in its Source Scripts list
-            Assert.Contains(scriptTextAsset, new List<TextAsset>(programImporter.sourceScripts));
+            // The project includes the script importer in its Source Scripts list
+            Assert.Contains(scriptTextAsset, new List<TextAsset>(projectImporter.sourceScripts));
         }
     }
 }